pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["examples"]
where = ["./"]

[project]
name = "dbrxish"
version = "0.0.1"
description = "Simple module to set up a Databricks notebook session in your local development environment.  Currently only works for 15.4 LTS runtimes."
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
keywords = ["databricks", "session", "local"]
authors = [
  {name = "Stephenson, Shane", email = "stephenson.shane.a@gmail.com" }
]
maintainers = [
  {name = "Stephenson, Shane", email = "stephenson.shane.a@gmail.com" }
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
  "databricks-connect==15.4",
  "databricks-sdk==0.33.0",
  "ipython"
]

[project.optional-dependencies]
dev = ["ruff", "wheel"]

[project.urls]
"Homepage" = "https://github.com/shaneish/databricks_sessionizer"
